+++ packages/web-app/app/_providers/idos/config.ts
@@ -1,0 +1,9 @@
+export const idOsConfig = {
+  container: '#idos',
+  nodeUrl: process.env.NEXT_PUBLIC_IDOS_NODE_URL,
+  dbId: process.env.NEXT_PUBLIC_IDOS_DB_ID,
+  evmGrantsOptions: {
+    contractAddress: process.env.NEXT_PUBLIC_IDOS_CONTRACT_ADDRESS,
+    chainId: process.env.NEXT_PUBLIC_IDOS_CHAIN_ID,
+  },
+};
